Delay Flip Flop
The D flip-flop is a two-input flip-flop used to store data at a predetermined time and hold it until it is needed. It is sometimes called a delay flip-flop because the data input is delayed up to one clock pulse before it is seen in the output. The clock is a timing pulse generated by the equipment to control operations