JK-Flip Flop
The JK flip-flop can be used to store temporary data in digital systems. The JK flip-flop can be used to synchronize data signals between two digital circuits, ensuring that they are operating on the same clock cycle. This makes it useful in applications where timing is critical. Designers usually prefer SR Flip Flop or JK Master Slave for design because of a drawback of timing problem known as "RACE". The condition of RACE arises if the output Q changes its state before the timing pulse of the clock input has time to go in the OFF state.